more testing and homework solution
